Continued code cleanup.
Show the master to slave and slave to master values in the statistics
output.

Add a Doxyfile and a tags target for building working cross references.
   NOTE: This requires cscope, global, doxygen and dot to be installed

Change gettimeofday() to clock_gettime() which gets us a timespec
and is therefore closer to what we want.  No more multiply/divide
by 1000 in the code.


git-svn-id: https://ptpd.svn.sourceforge.net/svnroot/ptpd/trunk@76 6eaf9dc1-1f19-0410-a618-e4a8c8cabde7