Add code to handle leap seconds properly.
In order to properly handle announced leap seconds at run time
we have to take an action when we see them, and not simply
set the flags.  This code handles that case and should
be thoroughly tested prior to the next leap second,
which occurs on June 30th 2012.
Submitted by:	Wojciech Owczarek


git-svn-id: https://ptpd.svn.sourceforge.net/svnroot/ptpd/trunk@203 6eaf9dc1-1f19-0410-a618-e4a8c8cabde7